requirements.txt
Django
pyTelegramBotAPI
djangorestframework
python-dotenv